+++ packages/i18n/src/formatters.ts
@@ -1,0 +1,8 @@
+/**
+ * Given a string, convert it to sentence case.
+ * @param item - The string to sentence case
+ * @returns
+ */
+export function sentence(str: string): string {
+  return str[0].toUpperCase() + str.substr(1)
+}