+++ packages/validation/src/index.ts
@@ -7,4 +7,5 @@ export {
   FormKitValidationIntent,
   FormKitValidationRule,
   FormKitValidationRules,
+  FormKitValidationMessages,
 } from './validation'